Used BigDecimal for all datatypes to avoid losing precision (#34)
* Increased the column length market_state.ma_type

* moved the services to download data and compute metrics to the "core" package along with the TechnicalAnalysisEngine.

- Used Numeric data type everywhere in db layer
- Used BigDecimal everywhere in app layer
- Manual string to big decimal calculation from the CSV
- Used only two crypto for testing BTC and ETH

+++ backend/src/main/java/com/prem/ta/models/OHLCVMetrics.java
@@ -1,11 +1,13 @@
 package com.prem.ta.models;
 
+import java.math.BigDecimal;
+
 public record OHLCVMetrics(
-        double open,
-        double high,
-        double low,
-        double close,
-        double volume,
-        double vwap
+        BigDecimal open,
+        BigDecimal high,
+        BigDecimal low,
+        BigDecimal close,
+        BigDecimal volume,
+        BigDecimal vwap
 ) {
 }
